add differential tests against evm
