+++ Cargo.toml
@@ -45,3 +45,8 @@ git = "https://github.com/TheDan64/inkwell.git"
 commit = "d916c66"
 default-features = false
 features = ["serde", "llvm18-0", "no-libffi-linking", "target-riscv"]
+
+[profile.benchmark]
+inherits = "release"
+lto = true
+codegen-units = 1
