+++ crates/benchmarks/EXECUTE.md
@@ -4,66 +4,14 @@
 
 - [Benchmark Results](#benchmark-results)
     - [Baseline](#baseline)
-    - [OddProduct](#oddproduct)
-    - [TriangleNumber](#trianglenumber)
-    - [FibonacciRecursive](#fibonaccirecursive)
-    - [FibonacciIterative](#fibonacciiterative)
-    - [FibonacciBinet](#fibonaccibinet)
-    - [SHA1](#sha1)
 
 ## Benchmark Results
 
 ### Baseline
 
-|         | `EVM`                     | `PVMInterpreter`                 | `PVM`                             |
-|:--------|:--------------------------|:---------------------------------|:--------------------------------- |
-| **`0`** | `901.74 ns` (✅ **1.00x**) | `736.09 ns` (✅ **1.23x faster**) | `26.38 us` (❌ *29.26x slower*)    |
-
-### OddProduct
-
-|              | `EVM`                    | `PVMInterpreter`                | `PVM`                              |
-|:-------------|:-------------------------|:--------------------------------|:---------------------------------- |
-| **`10000`**  | `7.45 ms` (✅ **1.00x**)  | `5.46 ms` (✅ **1.36x faster**)  | `136.31 us` (🚀 **54.67x faster**)  |
-| **`100000`** | `75.36 ms` (✅ **1.00x**) | `53.81 ms` (✅ **1.40x faster**) | `885.12 us` (🚀 **85.15x faster**)  |
-
-### TriangleNumber
-
-|              | `EVM`                    | `PVMInterpreter`                | `PVM`                              |
-|:-------------|:-------------------------|:--------------------------------|:---------------------------------- |
-| **`10000`**  | `5.15 ms` (✅ **1.00x**)  | `4.03 ms` (✅ **1.28x faster**)  | `117.30 us` (🚀 **43.93x faster**)  |
-| **`100000`** | `50.90 ms` (✅ **1.00x**) | `42.74 ms` (✅ **1.19x faster**) | `741.62 us` (🚀 **68.63x faster**)  |
-
-### FibonacciRecursive
-
-|          | `EVM`                     | `PVMInterpreter`                 | `PVM`                              |
-|:---------|:--------------------------|:---------------------------------|:---------------------------------- |
-| **`12`** | `258.01 us` (✅ **1.00x**) | `601.14 us` (❌ *2.33x slower*)   | `41.21 us` (🚀 **6.26x faster**)    |
-| **`16`** | `1.71 ms` (✅ **1.00x**)   | `3.96 ms` (❌ *2.32x slower*)     | `103.56 us` (🚀 **16.48x faster**)  |
-| **`20`** | `11.89 ms` (✅ **1.00x**)  | `28.21 ms` (❌ *2.37x slower*)    | `421.98 us` (🚀 **28.19x faster**)  |
-
-### FibonacciIterative
-
-|           | `EVM`                     | `PVMInterpreter`                 | `PVM`                            |
-|:----------|:--------------------------|:---------------------------------|:-------------------------------- |
-| **`64`**  | `26.84 us` (✅ **1.00x**)  | `110.57 us` (❌ *4.12x slower*)   | `40.37 us` (❌ *1.50x slower*)    |
-| **`128`** | `53.12 us` (✅ **1.00x**)  | `213.22 us` (❌ *4.01x slower*)   | `41.64 us` (✅ **1.28x faster**)  |
-| **`256`** | `110.52 us` (✅ **1.00x**) | `442.46 us` (❌ *4.00x slower*)   | `45.43 us` (🚀 **2.43x faster**)  |
-
-### FibonacciBinet
-
-|           | `EVM`                    | `PVMInterpreter`                 | `PVM`                            |
-|:----------|:-------------------------|:---------------------------------|:-------------------------------- |
-| **`64`**  | `20.10 us` (✅ **1.00x**) | `131.77 us` (❌ *6.55x slower*)   | `34.48 us` (❌ *1.72x slower*)    |
-| **`128`** | `23.61 us` (✅ **1.00x**) | `149.70 us` (❌ *6.34x slower*)   | `36.72 us` (❌ *1.56x slower*)    |
-| **`256`** | `25.94 us` (✅ **1.00x**) | `178.89 us` (❌ *6.90x slower*)   | `36.53 us` (❌ *1.41x slower*)    |
-
-### SHA1
-
-|           | `EVM`                     | `PVMInterpreter`                 | `PVM`                             |
-|:----------|:--------------------------|:---------------------------------|:--------------------------------- |
-| **`1`**   | `226.79 us` (✅ **1.00x**) | `340.44 us` (❌ *1.50x slower*)   | `38.29 us` (🚀 **5.92x faster**)   |
-| **`64`**  | `453.84 us` (✅ **1.00x**) | `573.04 us` (❌ *1.26x slower*)   | `46.06 us` (🚀 **9.85x faster**)   |
-| **`256`** | `1.12 ms` (✅ **1.00x**)   | `1.42 ms` (❌ *1.27x slower*)     | `68.83 us` (🚀 **16.30x faster**)  |
+|         | `EVM`                     | `PVMInterpreter`                  |
+|:--------|:--------------------------|:--------------------------------- |
+| **`0`** | `533.12 ns` (✅ **1.00x**) | `471.42 ns` (✅ **1.13x faster**)  |
 
 ---
 Made with [criterion-table](https://github.com/nu11ptr/criterion-table)
