+++ crates/benchmarks/src/runtimes/mod.rs
@@ -1,3 +1,1 @@
-#[cfg(feature = "bench-evm")]
-pub mod evm;
 pub mod polkavm;
