+++ crates/differential/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "revive-differential"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+evm-interpreter = { workspace = true }
+primitive-types = { workspace = true }
