+++ crates/integration/Cargo.toml
@@ -9,10 +9,12 @@ edition = "2021"
 polkavm = { workspace = true }
 alloy-primitives = { workspace = true }
 alloy-sol-types = { workspace = true }
-revive-solidity = { path = "../solidity" }
-era-compiler-llvm-context = { path = "../llvm-context" }
 hex = { workspace = true }
 env_logger = { workspace = true }
+
+revive-solidity = { path = "../solidity" }
+revive-differential = { path = "../differential" }
+era-compiler-llvm-context = { path = "../llvm-context" }
 
 [dev-dependencies]
 sha1 = { workspace = true }
